@@ -1,3 +1,4 @@
+#if os(macOS)
 import AppKit.NSMenu
 
 /**
@@ -364,7 +365,7 @@ public enum KeyboardShortcuts {
 	You can safely call this even if the user has not yet set a keyboard shortcut. It will just be inactive until they do.
 
 	```swift
-	import Cocoa
+	import AppKit
 	import KeyboardShortcuts
 
 	@main
@@ -390,7 +391,7 @@ public enum KeyboardShortcuts {
 	You can safely call this even if the user has not yet set a keyboard shortcut. It will just be inactive until they do.
 
 	```swift
-	import Cocoa
+	import AppKit
 	import KeyboardShortcuts
 
 	@main
@@ -458,8 +459,7 @@ public enum KeyboardShortcuts {
 }
 
 extension KeyboardShortcuts {
-	@available(macOS 10.15, *)
-	public enum EventType {
+	public enum EventType: Sendable {
 		case keyDown
 		case keyUp
 	}
@@ -493,7 +493,6 @@ extension KeyboardShortcuts {
 
 	- Note: This method is not affected by `.removeAllHandlers()`.
 	*/
-	@available(macOS 10.15, *)
 	public static func events(for name: Name) -> AsyncStream<KeyboardShortcuts.EventType> {
 		AsyncStream { continuation in
 			let id = UUID()
@@ -550,44 +549,12 @@ extension KeyboardShortcuts {
 
 	- Note: This method is not affected by `.removeAllHandlers()`.
 	*/
-	@available(macOS 10.15, *)
 	public static func events(_ type: EventType, for name: Name) -> AsyncFilterSequence<AsyncStream<EventType>> {
 		events(for: name).filter { $0 == type }
-	}
-
-	@available(macOS 10.15, *)
-	@available(*, deprecated, renamed: "events(_:for:)")
-	public static func on(_ type: EventType, for name: Name) -> AsyncStream<Void> {
-		AsyncStream { continuation in
-			let id = UUID()
-
-			switch type {
-			case .keyDown:
-				streamKeyDownHandlers[name, default: [:]][id] = {
-					continuation.yield()
-				}
-			case .keyUp:
-				streamKeyUpHandlers[name, default: [:]][id] = {
-					continuation.yield()
-				}
-			}
-
-			registerShortcutIfNeeded(for: name)
-
-			continuation.onTermination = { _ in
-				switch type {
-				case .keyDown:
-					streamKeyDownHandlers[name]?[id] = nil
-				case .keyUp:
-					streamKeyUpHandlers[name]?[id] = nil
-				}
-
-				unregisterShortcutIfNeeded(for: name)
-			}
-		}
 	}
 }
 
 extension Notification.Name {
 	static let shortcutByNameDidChange = Self("KeyboardShortcuts_shortcutByNameDidChange")
 }
+#endif
