@@ -1,3 +1,4 @@
+#if os(macOS)
 import AppKit
 import Carbon.HIToolbox
 
@@ -5,7 +6,7 @@ extension KeyboardShortcuts {
 	/**
 	A keyboard shortcut.
 	*/
-	public struct Shortcut: Hashable, Codable {
+	public struct Shortcut: Hashable, Codable, Sendable {
 		/**
 		Carbon modifiers are not always stored as the same number.
 
@@ -113,6 +114,7 @@ extension KeyboardShortcuts.Shortcut {
 	/**
 	Recursively finds a menu item in the given menu that has a matching key equivalent and modifier.
 	*/
+	@MainActor
 	func menuItemWithMatchingShortcut(in menu: NSMenu) -> NSMenuItem? {
 		for item in menu.items {
 			var keyEquivalent = item.keyEquivalent
@@ -144,6 +146,7 @@ extension KeyboardShortcuts.Shortcut {
 	/**
 	Returns a menu item in the app's main menu that has a matching key equivalent and modifier.
 	*/
+	@MainActor
 	var takenByMainMenu: NSMenuItem? {
 		guard let mainMenu = NSApp.mainMenu else {
 			return nil
@@ -153,7 +156,7 @@ extension KeyboardShortcuts.Shortcut {
 	}
 }
 
-private var keyToCharacterMapping: [KeyboardShortcuts.Key: String] = [
+private let keyToCharacterMapping: [KeyboardShortcuts.Key: String] = [
 	.return: "↩",
 	.delete: "⌫",
 	.deleteForward: "⌦",
@@ -161,7 +164,7 @@ private var keyToCharacterMapping: [KeyboardShortcuts.Key: String] = [
 	.escape: "⎋",
 	.help: "?⃝",
 	.home: "↖",
-	.space: "Space", // This matches what macOS uses.
+	.space: "space_key".localized, // This matches what macOS uses.
 	.tab: "⇥",
 	.pageUp: "⇞",
 	.pageDown: "⇟",
@@ -223,7 +226,7 @@ private func stringFromKeyCode(_ keyCode: Int) -> String {
 	String(format: "%C", keyCode)
 }
 
-private var keyToKeyEquivalentString: [KeyboardShortcuts.Key: String] = [
+private let keyToKeyEquivalentString: [KeyboardShortcuts.Key: String] = [
 	.space: stringFromKeyCode(0x20),
 	.f1: stringFromKeyCode(NSF1FunctionKey),
 	.f2: stringFromKeyCode(NSF2FunctionKey),
@@ -248,10 +251,8 @@ private var keyToKeyEquivalentString: [KeyboardShortcuts.Key: String] = [
 ]
 
 extension KeyboardShortcuts.Shortcut {
+	@MainActor // `TISGetInputSourceProperty` crashes if called on a non-main thread.
 	public func keyToCharacter() -> String? {
-		// `TISCopyCurrentASCIICapableKeyboardLayoutInputSource` works on a background thread, but crashes when used in a `NSBackgroundActivityScheduler` task, so we guard against that. It only crashes when running from Xcode, not in release builds, but it's probably safest to not call it from a `NSBackgroundActivityScheduler` no matter what.
-		assert(!DispatchQueue.isCurrentQueueNSBackgroundActivitySchedulerQueue, "This method cannot be used in a `NSBackgroundActivityScheduler` task")
-
 		// Some characters cannot be automatically translated.
 		if
 			let key,
@@ -300,6 +301,7 @@ extension KeyboardShortcuts.Shortcut {
 
 	- Note: Don't forget to also pass `.modifiers` to `NSMenuItem#keyEquivalentModifierMask`.
 	*/
+	@MainActor
 	var keyEquivalent: String {
 		let keyString = keyToCharacter() ?? ""
 
@@ -327,12 +329,14 @@ extension KeyboardShortcuts.Shortcut: CustomStringConvertible {
 	//=> "⌘A"
 	```
 	*/
+	@MainActor
 	public var description: String {
         if modifiers.contains(.function), key?.isFunctionKey == true {
             return modifiers.subtracting(.function).description + (keyToCharacter()?.capitalized ?? "�")
         } else {
-            // We use `.capitalized` so it correctly handles “⌘Space”.
-            return modifiers.description + (keyToCharacter()?.capitalized ?? "�")
+			// We use `.capitalized` so it correctly handles “⌘Space”.
+			modifiers.presentableDescription + (keyToCharacter()?.capitalized ?? "�")
         }
 	}
 }
+#endif
